refactor(project): Standardizes project structure and naming conventions.
Renames packages to follow a consistent snake_case naming convention, improving code organization and clarity.
