--- internal/git/git_test.go
@@ -1,5 +1,0 @@
-package git
-
-import "testing"
-
-func TestGit(t *testing.T) {}
