+++ internal/git_utils/git_utils.go
@@ -1,4 +1,4 @@
-package git
+package git_utils
 
 import (
 	"os/exec"
