+++ internal/git_utils/git_utils_test.go
@@ -1,0 +1,5 @@
+package git_utils
+
+import "testing"
+
+func TestGitUtils(t *testing.T) {}
